Ignore Errors that are not skip dir
This is for our own purposes - we almost always run full scans of the hdd, and
we often come across files and directories we cannot traverse or stat. In those
cases, we just ignore them.

In the future, we can inform the user we tried to stat a file and failed, but we
will continue regardless.